Auto-git/requirements.txt
gitpython==3.1.40
apscheduler==3.10.4
click==8.1.7
python-dateutil==2.8.2
colorama==0.4.6
